Move all the code to one file.
No need for three levels of python modules. This is a major change that
will break previous code.
