+++ fastapi_profiler/_version.py
@@ -1,2 +1,2 @@
-__version__ = "1.0.0"
+__version__ = "2.0.0"
 __author__ = "sunhailin-Leo"
