--- fastapi_profiler/profiler_middleware/__init__.py
@@ -1,1 +1,0 @@
-from ..profiler_middleware.profiler import PyInstrumentProfilerMiddleware
